requirements.txt
fastapi
uvicorn
python-dotenv
requests
pydantic
flask
pytest
playwright
chromium